Address feedback, move horizons designation to gem package.
