+++ modules/core/shared/src/main/scala/gem/HorizonsDesignation.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2017 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem.target
+package gem
 
 import gem.enum.HorizonsType
 import gem.parser.HorizonsDesignationParsers
