+++ modules/sql/src/main/resources/db/migration/V024__Horizons_Type.sql
@@ -4,17 +4,18 @@
 --
 
 CREATE TABLE e_horizons_type (
-    id   identifier            PRIMARY KEY,
-    name character varying(12) NOT NULL
+    id         identifier            PRIMARY KEY,
+    short_name character varying(12) NOT NULL,
+    long_name  character varying(21) NOT NULL
 );
 
 ALTER TABLE e_horizons_type OWNER TO postgres;
 
 
-COPY e_horizons_type (id, name) FROM stdin;
-Comet	Comet
-AsteroidNew	Asteroid New
-AsteroidOld	Asteroid Old
-MajorBody	Major Body
+COPY e_horizons_type (id, short_name, long_name) FROM stdin;
+Comet	Comet	Comet
+AsteroidNew	Asteroid New	Asteroid (New Format)
+AsteroidOld	Asteroid Old	Asteroid (Old Format)
+MajorBody	Major Body	Major Body
 \.
